Added 'Instructions to Get Started' section
I installed this plug-in last month but couldn't figure out how to get it to work. I came back to it a month later after having a lot more experience with Obsidian and figured it out. If I had 3 simple instructions, I could have avoided a lot of confusion. So I've added that to the readme. It's a quick-and-easy way to get started for new users!

Cheers! 

Also, changed Exclamation Marks in # Usage to :bangbang: emojis to attract more attention by using bright red color.